feat(frontend): audit log rework
